Move the layout-related code in `jni` to it's own file/module. Fixes #1104
