+++ crates/dc_jni/src/lib.rs
@@ -17,3 +17,4 @@ extern crate log;
 mod error;
 mod error_map;
 mod jni;
+mod layout_manager;
